requirements.txt
fastapi>=0.109.0,<1
uvicorn[standard]>=0.27.0,<1
pydantic>=2.5.0,<3
pydantic-settings>=2.1.0,<3
python-dotenv>=1.0.0,<2
sqlalchemy[asyncio]>=2.0.25,<3
asyncpg>=0.29.0,<1
alembic>=1.13.0,<2
psycopg2-binary>=2.9.9,<3
pgvector>=0.2.5,<0.4
boto3>=1.34.0,<2
openai>=1.40.0,<2
tiktoken>=0.7.0,<1
tenacity>=8.2.0,<10
httpx>=0.26.0,<1
sec-edgar-downloader>=5.0.0,<6
beautifulsoup4>=4.12.0,<5
lxml>=5.0.0,<6
langchain-text-splitters>=0.2.0,<0.4
pdfplumber>=0.11.0,<0.12
